+++ Source/Assimil8or/Preset/ZoneProperties.h
@@ -8,31 +8,31 @@ class ZoneProperties : public ValueTreeWrapper
 public:
     ZoneProperties () noexcept : ValueTreeWrapper (ZoneTypeId) {}
 
-    void setLevelOffset (float levelOffset, bool includeSelfCallback);
-    void setLoopLength (float loopLength, bool includeSelfCallback);
+    void setLevelOffset (double levelOffset, bool includeSelfCallback);
+    void setLoopLength (double loopLength, bool includeSelfCallback);
     void setLoopStart (int loopStart, bool includeSelfCallback);
-    void setMinVoltage (float minVoltage, bool includeSelfCallback);
-    void setPitchOffset (float pitchOffset, bool includeSelfCallback);
+    void setMinVoltage (double minVoltage, bool includeSelfCallback);
+    void setPitchOffset (double pitchOffset, bool includeSelfCallback);
     void setSample (juce::String sampleFileName, bool includeSelfCallback);
     void setSampleStart (int sampleStart, bool includeSelfCallback);
     void setSampleEnd (int sampleEnd, bool includeSelfCallback);
     void setSide (int side, bool includeSelfCallback);
 
-    float getLevelOffset ();
-    float getLoopLength ();
+    double getLevelOffset ();
+    double getLoopLength ();
     int getLoopStart ();
-    float getMinVoltage ();
-    float getPitchOffset ();
+    double getMinVoltage ();
+    double getPitchOffset ();
     juce::String getSample ();
     int getSampleStart ();
     int getSampleEnd ();
     int getSide ();
 
-    std::function<void (float levelOffset)> onLevelOffsetChange;
-    std::function<void (float loopLength)> onLoopLengthChange;
+    std::function<void (double levelOffset)> onLevelOffsetChange;
+    std::function<void (double loopLength)> onLoopLengthChange;
     std::function<void (int loopStart)> onLoopStartChange;
-    std::function<void (float minVoltage)> onMinVoltageChange;
-    std::function<void (float pitchOffset)> onPitchOffsetChange;
+    std::function<void (double minVoltage)> onMinVoltageChange;
+    std::function<void (double pitchOffset)> onPitchOffsetChange;
     std::function<void (juce::String sampleFileName)> onSampleChange;
     std::function<void (int sampleStart)> onSampleStartChange;
     std::function<void (int sampleEnd)> onSampleEndChange;
